Reorg hook 🪝
