+++ apps/ledger-live-mobile/src/screens/Nft/WalletNftGallery/index.tsx
@@ -1,18 +1,45 @@
-import React from "react";
+import React, { useMemo } from "react";
 import { Flex } from "@ledgerhq/native-ui";
 import { useSelector } from "react-redux";
 import { useTheme } from "styled-components/native";
 import NftList from "~/components/Nft/NftGallery/NftList";
 import NftGalleryEmptyState from "../NftGallery/NftGalleryEmptyState";
 import CollapsibleHeaderScrollView from "~/components/WalletTab/CollapsibleHeaderScrollView";
-import { hasNftsSelector } from "~/reducers/accounts";
+import { accountsSelector, filteredNftsSelector, hasNftsSelector } from "~/reducers/accounts";
 import { useNftGallery } from "~/hooks/useNftGallery";
+import { isEqual } from "lodash";
+import { galleryChainFiltersSelector } from "~/reducers/nft";
 
 const WalletNftGallery = () => {
   const { space } = useTheme();
   const hasNFTs = useSelector(hasNftsSelector);
+  const accounts = useSelector(accountsSelector);
 
-  const { isLoading, hasNextPage, error, fetchNextPage, refetch, parsedData } = useNftGallery();
+  const chainFilters = useSelector(galleryChainFiltersSelector);
+  const nftsOwned = useSelector(filteredNftsSelector, isEqual);
+
+  const addresses = useMemo(
+    () => [
+      ...new Set(
+        accounts.map(account => account.freshAddress).filter(addr => addr.startsWith("0x")),
+      ),
+    ],
+    [accounts],
+  );
+
+  const chains = useMemo(
+    () =>
+      Object.entries(chainFilters)
+        .filter(([_, value]) => value)
+        .map(([key, _]) => key),
+    [chainFilters],
+  );
+
+  const { isLoading, hasNextPage, error, fetchNextPage, refetch, parsedData } = useNftGallery({
+    addresses,
+    chains,
+    nftsOwned,
+  });
 
   return (
     <Flex flex={1} testID="wallet-nft-gallery-screen">
